Rename some lib files according to namespaces
