+++ lib/liquid/rails/template_handler.rb
